Added Abalone Snail Age Prediction Model
I have added an Abalone Age Prediction Model. The age of abalone is determined by cutting the shell through the cone, staining it, and counting the number of rings through a microscope. In order to make this process easier, the Machine Learning model can easily predict the age of the snails by analysing the data and applying regression algorithms.